Add files via upload
First Attempt at coding Camera Controls and PlayerMovement. So I chose a Character Controller which i will be referencing as CC from now on. The reason I chose CC over A RigidBody is because a CC has everything i want pretty much out of the box/ RigidBodys also dont feel very *snappy* with movement I want the player to stop moving when they realease a button, which is exactly what CC.Move() provides it doesnt feel like the player is sliding and Theres no extra code required to get that feeling. The camera also has a hard time following RigidBodies because rigidbodies work on the Physics loop which by default is 0.222 Seconds so about 50 FPS this is not a ideal framerate. For the reasons listed above i will be using a CC instead.